Convert F and t values to Noncentrality-Based Confidence Intervals

These methods searches for a the best ncp (non-central parameters) for either the t distribution (for r and Cohen's d effect sizes) of the F distribution (for the PVE effect sizes) for the desired tail-probabilities, and then convert these ncps to the corresponding effect sizes using effectsize's conversion functions.

F_to_PVE_CI(
  f,
  df,
  df_error,
  conf.level = 0.9,
  type = c("eta2", "epsilon2", "omega2")
)

t_to_PVE_CI(
  t,
  df_error,
  conf.level = 0.9,
  type = c("eta2", "epsilon2", "omega2")
)

t_to_r_CI(t, df_error, conf.level = 0.95)

F_to_r_CI(f, df, df_error, conf.level = 0.95)

t_to_d_CI(t, df_error, conf.level = 0.95, pooled = FALSE)

F_to_d_CI(f, df, df_error, conf.level = 0.95, pooled = FALSE)

Arguments

f, t
The t or the F statistics.

df, df_error
Degrees of freedom of numerator or of the error estimate (i.e., the residuals).

conf.level
Confidence level.

type
See the corresponding functions in effectsize.

Details

All effect sizes are converted using simple conversion formulas. See formulas for partial variance explained (PVE) and for d and r.

CI for PVE are based on Steiger (2004, eq. 16), and for Cohen's d are based on Cumming and Finch (2001, eq. 17-18). Read here for more.

References

Steiger, J. H. (2004). Beyond the F test: Effect size confidence intervals and tests of close fit in the analysis of variance and contrast analysis. Psychological Methods, 9, 164-182.

Cumming, G., & Finch, S. (2001). A primer on the understanding, use, and calculation of confidence intervals that are based on central and noncentral distributions. Educational and Psychological Measurement, 61(4), 532-574.

Examples

# CI to eta2
if (requireNamespace("MBESS", quietly = TRUE)) {
  cbind(MSBMisc = F_to_PVE_CI(4, 3, 123)[-2],
        MBESS = unlist(MBESS::ci.pvaf(
          4, 3, 123, N = 127, conf.level = 0.9
        ))[c(1, 3)])
}

#>          MSBMisc      MBESS
#> Lower 0.01351694 0.01309678
#> Upper 0.16330450 0.15897876

# CI for d
if (requireNamespace("MBESS", quietly = TRUE)) {
  cbind(MSBMisc = t_to_d_CI(4, 68)[-2],
        MBESS = unlist(MBESS::ci.smd(4, n.1 = 40, n.2 = 30))[-2])
}

#>         MSBMisc     MBESS
#> Lower 0.4644917 0.4625522
#> Upper 1.4693874 1.4632522

# CI for r
res <- cor.test(iris[[1]], iris[[2]])
cbind(MSBMisc = t_to_r_CI(res$statistic, res$parameter)[-2],
      stats = res$conf.int)

#>           MSBMisc       stats
#> Lower -0.27121237 -0.27269325
#> Upper  0.04373319  0.04351158