Sequence Generation Baed on Range

Sequence Generation Baed on Range

seq_range(x, length.out = 20, na.rm = TRUE)

mean_sd(x, na.rm = TRUE, out = c("vector", "data.frame"))

Arguments

x

A numeric vector

length.out

desired length of the sequence. A non-negative number, which for seq and seq.int will be rounded up if fractional.

na.rm

a logical value indicating whether NA values should be stripped before the computation proceeds.

out

If "data.frame" can be used as a summary function in ggplot2.

Examples

set.seed(1)
x <- rnorm(100)
seq_range(x, length.out = 5)
#> [1] -2.21469989 -1.06062048  0.09345894  1.24753835  2.40161776

mean_sd(x)
#>        -SD       Mean        +SD 
#> -0.7893120  0.1088874  1.0070867 

if (FALSE) {
library(ggplot2)
ggplot(mtcars, aes(cyl, mpg)) +
  stat_summary(fun.data = mean_sd, fun.args = list(out = "data.frame"))
}